Replace Fragment with Controller
