+++ build.gradle
@@ -10,7 +10,6 @@ buildscript {
         compileSdkVersion = 28
         buildToolsVersion = '28.0.1'
         supportLibraryVersion = '28.0.0-alpha3'
-        takisoftFixVersion = '27.1.1.2'
         roomVersion = '1.1.1'
         workVersion = '1.0.0-alpha04'
         junitVersion = '4.12'
@@ -21,9 +20,7 @@ buildscript {
     repositories {
         google()
         jcenter()
-        maven {
-            url 'https://maven.fabric.io/public'
-        }
+        maven { url 'https://maven.fabric.io/public' }
     }
 
     dependencies {
@@ -39,6 +36,7 @@ allprojects {
     repositories {
         google()
         jcenter()
+        maven { url 'https://jitpack.io' }
     }
 }
