--- mobile/src/main/java/com/github/shadowsocks/AboutFragment.kt
@@ -1,64 +1,0 @@
-/*******************************************************************************
- *                                                                             *
- *  Copyright (C) 2017 by Max Lv <max.c.lv@gmail.com>                          *
- *  Copyright (C) 2017 by Mygod Studio <contact-shadowsocks-android@mygod.be>  *
- *                                                                             *
- *  This program is free software: you can redistribute it and/or modify       *
- *  it under the terms of the GNU General Public License as published by       *
- *  the Free Software Foundation, either version 3 of the License, or          *
- *  (at your option) any later version.                                        *
- *                                                                             *
- *  This program is distributed in the hope that it will be useful,            *
- *  but WITHOUT ANY WARRANTY; without even the implied warranty of             *
- *  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the              *
- *  GNU General Public License for more details.                               *
- *                                                                             *
- *  You should have received a copy of the GNU General Public License          *
- *  along with this program. If not, see <http://www.gnu.org/licenses/>.       *
- *                                                                             *
- *******************************************************************************/
-
-package com.github.shadowsocks
-
-import android.content.Intent
-import android.net.Uri
-import android.os.Bundle
-import android.support.v4.text.HtmlCompat
-import android.text.SpannableStringBuilder
-import android.text.method.LinkMovementMethod
-import android.text.style.ClickableSpan
-import android.text.style.URLSpan
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import android.widget.TextView
-
-class AboutFragment : ToolbarFragment() {
-    override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?): View? =
-            inflater.inflate(R.layout.layout_about, container, false)
-
-    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
-        super.onViewCreated(view, savedInstanceState)
-        toolbar.title = getString(R.string.about_title, BuildConfig.VERSION_NAME)
-        view.findViewById<TextView>(R.id.tv_about).apply {
-            text = SpannableStringBuilder(HtmlCompat.fromHtml(
-                    resources.openRawResource(R.raw.about).bufferedReader().readText(),
-                    HtmlCompat.FROM_HTML_SEPARATOR_LINE_BREAK_LIST_ITEM)).apply {
-                for (span in getSpans(0, length, URLSpan::class.java)) {
-                    setSpan(object : ClickableSpan() {
-                        override fun onClick(view: View) {
-                            if (span.url.startsWith("mailto:")) {
-                                startActivity(Intent.createChooser(Intent().apply {
-                                    action = Intent.ACTION_SENDTO
-                                    data = Uri.parse(span.url)
-                                }, getString(R.string.send_email)))
-                            } else (activity as MainActivity).launchUrl(span.url)
-                        }
-                    }, getSpanStart(span), getSpanEnd(span), getSpanFlags(span))
-                    removeSpan(span)
-                }
-            }
-            movementMethod = LinkMovementMethod.getInstance()
-        }
-    }
-}
