+++ mobile/src/main/res/layout/layout_about.xml
@@ -1,20 +1,14 @@
 <?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    android:orientation="vertical"
+<ScrollView
+    xmlns:android="http://schemas.android.com/apk/res/android"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    android:id="@+id/about_container">
-    <include layout="@layout/toolbar_light_dark"/>
-    <ScrollView
+    android:scrollbars="vertical">
+    <com.github.shadowsocks.widget.HackyTextView
+        android:id="@+id/tv_about"
         android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:scrollbars="vertical">
-        <com.github.shadowsocks.widget.HackyTextView
-            android:id="@+id/tv_about"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:padding="16dp"
-            android:textIsSelectable="true"
-            android:textAppearance="@style/TextAppearance.AppCompat.Body1"/>
-    </ScrollView>
-</LinearLayout>
+        android:layout_height="wrap_content"
+        android:padding="16dp"
+        android:textIsSelectable="true"
+        android:textAppearance="@style/TextAppearance.AppCompat.Body1"/>
+</ScrollView>
