+++ mobile/src/main/res/layout/layout_list.xml
@@ -1,17 +1,18 @@
 <?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-              xmlns:tools="http://schemas.android.com/tools"
-              android:layout_width="match_parent"
-              android:layout_height="match_parent"
-              android:orientation="vertical">
-    <include layout="@layout/toolbar_light_dark"/>
-    <android.support.v7.widget.RecyclerView android:id="@+id/list"
-                                            tools:listitem="@layout/layout_profile"
-                                            android:layout_width="match_parent"
-                                            android:layout_height="0dp"
-                                            android:layout_weight="1"
-                                            android:paddingBottom="32dp"
-                                            android:scrollbars="vertical"
-                                            android:clipChildren="false"
-                                            android:clipToPadding="false"/>
+<LinearLayout
+    xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    android:orientation="vertical">
+    <android.support.v7.widget.RecyclerView
+        android:id="@+id/list"
+        tools:listitem="@layout/layout_profile"
+        android:layout_width="match_parent"
+        android:layout_height="0dp"
+        android:layout_weight="1"
+        android:paddingBottom="32dp"
+        android:scrollbars="vertical"
+        android:clipChildren="false"
+        android:clipToPadding="false"/>
 </LinearLayout>
