+++ mobile/src/main/res/layout/layout_profile_config.xml
@@ -1,12 +1,13 @@
 <?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-              android:orientation="vertical"
-              android:layout_width="match_parent"
-              android:layout_height="match_parent">
-    <include layout="@layout/toolbar_light_dark" />
-    <fragment android:id="@+id/content"
-              class="com.github.shadowsocks.ProfileConfigFragment"
-              android:layout_width="match_parent"
-              android:layout_height="0dp"
-              android:layout_weight="1"/>
+<LinearLayout
+    xmlns:android="http://schemas.android.com/apk/res/android"
+    android:orientation="vertical"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent">
+    <include layout="@layout/toolbar_light_dark"/>
+    <com.bluelinelabs.conductor.ChangeHandlerFrameLayout
+        android:id="@+id/controller_container"
+        android:background="?android:attr/windowBackground"
+        android:layout_width="match_parent"
+        android:layout_height="match_parent" />
 </LinearLayout>
