+++ plugin/build.gradle
@@ -24,9 +24,8 @@ android {
 }
 
 dependencies {
-    api "com.android.support:preference-v14:$supportLibraryVersion"
-    api "com.takisoft.fix:preference-v7:$takisoftFixVersion"
     api "org.jetbrains.kotlin:kotlin-stdlib-jdk7:$kotlinVersion"
+    api 'com.github.onlymash:conductor-preference:1.0.3'
     testImplementation "junit:junit:$junitVersion"
     androidTestImplementation "com.android.support.test:runner:$androidTestVersion"
     androidTestImplementation "com.android.support.test.espresso:espresso-core:$androidEspressoVersion"
