+++ plugin/src/main/res/layout/toolbar_light_dark.xml
@@ -1,12 +1,17 @@
 <?xml version="1.0" encoding="utf-8"?>
-<android.support.v7.widget.Toolbar
+<android.support.design.widget.AppBarLayout
     xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    android:layout_height="?attr/actionBarSize"
     android:layout_width="match_parent"
-    android:background="?attr/colorPrimary"
-    android:elevation="4dp"
-    android:touchscreenBlocksFocus="false"
-    android:theme="@style/ThemeOverlay.AppCompat.Dark.ActionBar"
-    app:popupTheme="@style/ThemeOverlay.AppCompat.DayNight"
-    android:id="@+id/toolbar" />
+    android:layout_height="wrap_content">
+    <android.support.v7.widget.Toolbar
+        android:layout_height="?attr/actionBarSize"
+        android:layout_width="match_parent"
+        android:background="?attr/colorPrimary"
+        android:elevation="4dp"
+        android:touchscreenBlocksFocus="false"
+        android:theme="@style/ThemeOverlay.AppCompat.Dark.ActionBar"
+        app:popupTheme="@style/ThemeOverlay.AppCompat.DayNight"
+        android:id="@+id/toolbar" />
+
+</android.support.design.widget.AppBarLayout>
